Add VB support
